Moved Spellchecker Decorator to one file and converted to ES6 class
