+++ scripts/core/editor3/components/spellchecker/SpellcheckerError.jsx
@@ -1,5 +1,6 @@
-import React from 'react';
+import React, {Component} from 'react';
 import {connect} from 'react-redux';
+import ng from 'core/services/ng';
 import * as actions from '../../actions';
 
 /**
@@ -11,26 +12,52 @@ import * as actions from '../../actions';
  * @description The words with spellcheck errors are enclosed in this component in order to highlight
  * the error and allow the opening of the contextual spellchecker menu.
  */
+export class SpellcheckerErrorComponent extends Component {
+    static getDecorators() {
+        return [{
+            strategy: spellcheckStrategy,
+            component: SpellcheckerError
+        }];
+    }
 
-export const SpellcheckerErrorComponent = ({children, showContextMenu}) => {
-    const word = {
-        text: children[0].props.text,
-        offset: children[0].props.start
-    };
+    constructor(props) {
+        super(props);
 
-    const onContextMenu = (e) => {
-        const position = {
-            left: e.clientX - 20,
-            top: e.clientY + 10
-        };
+        this.onContextMenu = this.onContextMenu.bind(this);
+    }
 
+    onContextMenu(e) {
         e.preventDefault();
 
-        showContextMenu({word, position});
-    };
+        const {children, showContextMenu} = this.props;
+        const word = {text: children[0].props.text, offset: children[0].props.start};
+        const position = {left: e.clientX - 20, top: e.clientY + 10};
 
-    return <span className="word-typo" onContextMenu={onContextMenu}>{children}</span>;
-};
+        showContextMenu({word, position});
+    }
+
+    render() {
+        return <span className="word-typo" onContextMenu={this.onContextMenu}>{this.props.children}</span>;
+    }
+}
+
+/**
+ * @description For a block check the words that has errors
+ */
+function spellcheckStrategy(contentBlock, callback) {
+    const spellcheck = ng.get('spellcheck');
+    const WORD_REGEXP = /[0-9a-zA-Z\u00C0-\u1FFF\u2C00-\uD7FF]+/g;
+    const text = contentBlock.getText();
+
+    let matchArr, start, regex = WORD_REGEXP;
+
+    while ((matchArr = regex.exec(text)) !== null) {
+        start = matchArr.index;
+        if (!spellcheck.isCorrectWord(matchArr[0])) {
+            callback(start, start + matchArr[0].length);
+        }
+    }
+}
 
 
 /** Set the types of props for the spellchecker error component*/
